+++ admin/settings-page.php
@@ -99,7 +99,6 @@
                         </h3>
 				        <div class="inside fbak-inside">
                             <?php do_settings_sections('fbak_plugin_woo_option'); ?>
-                            <br><b><?php _e( 'Note:', 'fb-account-kit-login' ); ?></b> <i><?php _e( 'Please re-save your permalinks from <strong>Settings > Permalinks</strong> if you make any changes in Account Kit Authentication Endpoint URL.', 'fb-account-kit-login' ); ?></i>
                             <?php submit_button( __( 'Save Settings', 'fb-account-kit-login' ), 'primary save-settings', 'submit-woo' ); ?>
                         </div>
                     </div>
@@ -176,9 +175,22 @@
                                     $('#saveMessage').show().delay(4000).fadeOut();
                                     $(".save-settings").removeClass("disabled");
                                     $(".save-settings").val("<?php _e( 'Save Settings', 'fb-account-kit-login' ); ?>");
+                                    if ($('#changetrigger').val() == 'yes') {
+                                        $.ajax({
+                                            type: "POST",
+                                            url: "<?php echo admin_url( 'admin-ajax.php' ); ?>",
+                                            dataType: "json",
+                                            data: {
+                                                action: "fbak_trigger_flush_rewrite_rules",
+                                            }
+                                        });
+                                    }
                                 }
                             });
                             return false;
+                        });
+                        $("#fbak-woo-ep").change(function() {
+					        $('#changetrigger').val('yes');
                         });
                     });
                 </script>