+++ includes/functions.php
@@ -28,7 +28,7 @@ function fbak_guess_username_by_phone( $phone ) {
 
     // Ensure username is unique.
     $append = 1;
-    $o_username = $username;
+    $o_username = $username . '_';
 
     while ( username_exists( $username ) ) {
         $username = $o_username . $append;