+++ includes/js/profile.js
@@ -26,10 +26,6 @@
                 window.location.reload();
             });
         }
-        else if (response.status === "BAD_PARAMS") {
-            // handle bad parameters
-            alert( FBAccountKitLogin.bad_params );
-        }
     }
 
     // phone form submission handler